+++ crates/goose/src/agents/mod.rs
@@ -11,7 +11,7 @@ mod reply_parts;
 mod router_tool_selector;
 mod router_tools;
 mod schedule_tool;
-pub mod sub_agent_execution_tool;
+pub mod sub_recipe_execution_tool;
 pub mod sub_recipe_manager;
 pub mod subagent;
 pub mod subagent_handler;
