+++ crates/goose/src/agents/sub_recipe_execution_tool/executor.rs
@@ -3,11 +3,11 @@ use std::sync::Arc;
 use tokio::sync::mpsc;
 use tokio::time::Instant;
 
-use crate::agents::sub_agent_execution_tool::lib::{
+use crate::agents::sub_recipe_execution_tool::lib::{
     Config, ExecutionResponse, ExecutionStats, Task, TaskResult,
 };
-use crate::agents::sub_agent_execution_tool::tasks::process_task;
-use crate::agents::sub_agent_execution_tool::workers::{run_scaler, spawn_worker, SharedState};
+use crate::agents::sub_recipe_execution_tool::tasks::process_task;
+use crate::agents::sub_recipe_execution_tool::workers::{run_scaler, spawn_worker, SharedState};
 
 pub async fn execute_single_task(task: &Task, config: Config) -> ExecutionResponse {
     let start_time = Instant::now();
