+++ crates/goose/src/agents/sub_recipe_execution_tool/mod.rs
@@ -1,6 +1,6 @@
 mod executor;
 pub mod lib;
-pub mod sub_agent_execute_task_tool;
+pub mod sub_recipe_execute_task_tool;
 mod tasks;
 mod types;
 mod workers;
