+++ crates/goose/src/agents/sub_recipe_execution_tool/sub_recipe_execute_task_tool.rs
@@ -1,12 +1,14 @@
 use mcp_core::{tool::ToolAnnotations, Content, Tool, ToolError};
 use serde_json::Value;
 
-use crate::agents::{sub_agent_execution_tool::lib::execute_tasks, tool_execution::ToolCallResult};
+use crate::agents::{
+    sub_recipe_execution_tool::lib::execute_tasks, tool_execution::ToolCallResult,
+};
 
-pub const SUB_AGENT_EXECUTE_TASK_TOOL_NAME: &str = "sub_recipe__execute_task";
-pub fn create_sub_agent_execute_task_tool() -> Tool {
+pub const SUB_RECIPE_EXECUTE_TASK_TOOL_NAME: &str = "sub_recipe__execute_task";
+pub fn create_sub_recipe_execute_task_tool() -> Tool {
     Tool::new(
-        SUB_AGENT_EXECUTE_TASK_TOOL_NAME,
+        SUB_RECIPE_EXECUTE_TASK_TOOL_NAME,
         "Only use this tool when you want to execute sub recipe task. **DO NOT** use this tool when you want to execute sub agent task.   
         If the tasks are not specified to be executed in parallel, you should use this tool to run each task immediately by passing a single task to the tool for each run.
         If you want to execute tasks in parallel, you should pass a list of tasks to the tool.",
