+++ crates/goose/src/agents/sub_recipe_execution_tool/workers.rs
@@ -1,5 +1,5 @@
-use crate::agents::sub_agent_execution_tool::tasks::process_task;
-use crate::agents::sub_agent_execution_tool::types::{Task, TaskResult};
+use crate::agents::sub_recipe_execution_tool::tasks::process_task;
+use crate::agents::sub_recipe_execution_tool::types::{Task, TaskResult};
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 use std::sync::Arc;
 use tokio::sync::mpsc;
@@ -8,7 +8,7 @@ use tokio::time::{sleep, Duration};
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::agents::sub_agent_execution_tool::types::Task;
+    use crate::agents::sub_recipe_execution_tool::types::Task;
 
     #[tokio::test]
     async fn test_spawn_worker_returns_handle() {
